# Changes to support execution hooks