… src/test/java/rx/plugins/RxJavaPluginsTest.java