wrap the onSubscribe/Operator as soon as possible in lift.
wrap the onSubscribe in the constructor to get subclasses.
add code to the scalar synchronous observable to detect exec hook usage.