@akarnokd 2015-06-17
I'm willing to accept these changes but it is going to conflict with #2928.