@abersnaze 2015-06-18
I added a bunch of unit tests to show it was working as well. I had to do some funny things to the hook registration to make the test not so flaky.  @benjchristensen was concerned that the performance of would suffer.  I'm still working out that part.

Before
```
Benchmark                                 (size)   Mode   Samples        Score  Score error    Units
r.o.OperatorMapPerf.mapPassThru                1  thrpt         5 20861710.150  2528216.623    ops/s
r.o.OperatorMapPerf.mapPassThru             1000  thrpt         5   144703.043    16183.258    ops/s
r.o.OperatorMapPerf.mapPassThru          1000000  thrpt         5      144.271        8.701    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift         1  thrpt         5 22885527.107   995464.251    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift      1000  thrpt         5   145704.353     9729.424    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift   1000000  thrpt         5      138.445       17.020    ops/s
Benchmark                                          (size)   Mode   Samples        Score  Score error    Units
r.o.OperatorMergePerf.merge1SyncStreamOfN               1  thrpt         5  4839482.480   239709.614    ops/s
r.o.OperatorMergePerf.merge1SyncStreamOfN            1000  thrpt         5    28336.689     2016.760    ops/s
r.o.OperatorMergePerf.merge1SyncStreamOfN         1000000  thrpt         5       33.909        0.547    ops/s
r.o.OperatorMergePerf.mergeNAsyncStreamsOfN             1  thrpt         5    97959.393     4014.184    ops/s
r.o.OperatorMergePerf.mergeNAsyncStreamsOfN          1000  thrpt         5        4.924        1.077    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1              1  thrpt         5  4437187.814   167833.354    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1            100  thrpt         5   351315.999    44526.813    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1           1000  thrpt         5    37121.316     3895.097    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOfN              1  thrpt         5  5021914.272   412909.787    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOfN           1000  thrpt         5       32.252        5.023    ops/s
r.o.OperatorMergePerf.mergeTwoAsyncStreamsOfN           1  thrpt         5    76855.583      958.621    ops/s
r.o.OperatorMergePerf.mergeTwoAsyncStreamsOfN        1000  thrpt         5     1717.848      855.387    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1         1  thrpt         5  4643618.830   215086.259    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1      1000  thrpt         5    24258.258      612.329    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1   1000000  thrpt         5       30.282        1.005    ops/s
```

After
```
Benchmark                                 (size)   Mode   Samples        Score  Score error    Units
r.o.OperatorMapPerf.mapPassThru                1  thrpt         5 21350752.203  1037602.046    ops/s
r.o.OperatorMapPerf.mapPassThru             1000  thrpt         5   146674.940     5320.446    ops/s
r.o.OperatorMapPerf.mapPassThru          1000000  thrpt         5      146.099        2.902    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift         1  thrpt         5 22679101.907   333070.139    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift      1000  thrpt         5   144797.255     6644.360    ops/s
r.o.OperatorMapPerf.mapPassThruViaLift   1000000  thrpt         5      146.210        6.251    ops/s
Benchmark                                          (size)   Mode   Samples        Score  Score error    Units
r.o.OperatorMergePerf.merge1SyncStreamOfN               1  thrpt         5  4415320.588   156926.619    ops/s
r.o.OperatorMergePerf.merge1SyncStreamOfN            1000  thrpt         5    28327.275      953.037    ops/s
r.o.OperatorMergePerf.merge1SyncStreamOfN         1000000  thrpt         5       32.374        1.956    ops/s
r.o.OperatorMergePerf.mergeNAsyncStreamsOfN             1  thrpt         5    92374.205     4680.405    ops/s
r.o.OperatorMergePerf.mergeNAsyncStreamsOfN          1000  thrpt         5        4.881        0.489    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1              1  thrpt         5  4486377.199    73753.728    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1            100  thrpt         5   352662.487    14879.412    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOf1           1000  thrpt         5    46931.354     2035.109    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOfN              1  thrpt         5  4788335.318   487569.791    ops/s
r.o.OperatorMergePerf.mergeNSyncStreamsOfN           1000  thrpt         5       32.847        1.320    ops/s
r.o.OperatorMergePerf.mergeTwoAsyncStreamsOfN           1  thrpt         5    77044.016      501.663    ops/s
r.o.OperatorMergePerf.mergeTwoAsyncStreamsOfN        1000  thrpt         5     1737.993      751.681    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1         1  thrpt         5  4630445.306    39953.022    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1      1000  thrpt         5    23478.427     2397.959    ops/s
r.o.OperatorMergePerf.oneStreamOfNthatMergesIn1   1000000  thrpt         5       27.596        1.443    ops/s
```