@benjchristensen 2015-07-14
Merge conflicts, so closing. Fix up and resubmit.